Move NetworkManager partial classes to individual files.
