+++ Core/NetworkManager_Helpers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fba0fe21359df42078109dd6eccc8721
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
