+++ Core/NetworkManager_Inspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f75055fd7197d3289bccbb4e08bb2ebb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
